cuad-contract-analysis/requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
openai>=1.0.0
anthropic>=0.7.0
nltk>=3.8.0
regex>=2023.10.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
tenacity>=8.2.0
jupyter>=1.0.0
ipykernel>=6.25.0
plotly>=5.0.0
matplotlib>=3.0.0
Pillow>=9.0.0